Remove requirements.txt
The package specifications in `requirements.txt` are covered by
`setup.cfg`, but are also a bit out of sync.  (`case-utils` requires a
more recent RDFLib version, and `setup.cfg` requires a more specific
version of `case-utils`.)

I also did not see anywhere in the CI scripting where `requirements.txt`
was brought in for e.g. virtual environment setup.

Thus, this patch is filed to remove the potential for package version
conflicts.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>