Use local_uuid, not demo_uuid
A shortcoming in the documentation for the `case_utils.local_uuid`
module may have made it unclear what function to call.

`demo_uuid()` is effectively an internal function, and will raise a
runtime error for users who don't set the intentionally-awkward
environment variable to opt in to its behavior.

This patch switches use to the intended "public" interface of the
`case_utils.local_uuid` module.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>